// src/types/setting-Types.ts
export interface User {
  _id: string
  email: string
  phone_Number: string
  role: string
  address: any[] // Define a more specific type if needed
  __v: number
  ticketsAssigned?: any[] // Optional
  cartId?: string | null // Optional
  fcmToken?: string | null // Optional
  wishlistId?: string | null // Optional
  last_login?: string // Optional
  vehicle_details?: any[] // Optional
}
// ...
export interface UpdateUserPermissionsRequest {
  module: string
  role: string
  userId: string | User
  permissions: {
    allowedFields: string[]
    read: boolean
    write: boolean
    update: boolean
    delete: boolean
  }
  updatedBy?: string | User
}